Skip Mux delete when other local assets still reference it
Guard remote Mux asset deletion in DeleteMuxAsset so that an Asset-keyed
delete is a no-op when other local assets still reference the same
mux_id. The string-id path stays unchecked as the explicit escape
hatch (used by CreateMuxAsset for re-upload cleanup).

Extract the lookup to MirrorField::assetsByMuxId() so both the create
(pre-check) and delete (guard) paths share one implementation.
